+++ src/models/actions/ReorderAssetsAction.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace davidhirtz\yii2\cms\models\actions;
+
+use DateTime;
+use davidhirtz\yii2\cms\models\Entry;
+use davidhirtz\yii2\cms\models\Section;
+use davidhirtz\yii2\skeleton\models\Trail;
+use Yii;
+
+class ReorderAssetsAction extends ReorderActiveRecordsAction
+{
+    public function __construct(protected Entry|Section $parent, array $assetIds = [])
+    {
+        $assets = $parent->getAssets()
+            ->select(['id', 'position'])
+            ->andWhere(['id' => $assetIds])
+            ->orderBy(['position' => SORT_ASC])
+            ->all();
+
+        $order = array_flip($assetIds);
+
+        parent::__construct($assets, $order);
+    }
+
+    protected function afterReorder(): void
+    {
+        $trail = Trail::createOrderTrail($this->parent, Yii::t('cms', 'Asset order changed'));
+
+        $this->parent->updated_at = new DateTime();
+        $this->parent->update();
+
+        if($this->parent instanceof Section) {
+            $entry = $this->parent->entry;
+            Trail::createOrderTrail($entry, Yii::t('cms', 'Section asset order changed'), [
+                'trail_id' => $trail->id,
+            ]);
+
+            $entry->updated_at = new DateTime();
+            $entry->update();
+        }
+
+        parent::afterReorder();
+    }
+}