+++ src/modules/admin/controllers/SectionController.php
@@ -11,6 +11,7 @@ use davidhirtz\yii2\cms\modules\admin\data\EntryActiveDataProvider;
 use davidhirtz\yii2\cms\modules\ModuleTrait;
 use davidhirtz\yii2\cms\models\Section;
 use davidhirtz\yii2\cms\models\Entry;
+use davidhirtz\yii2\cms\models\actions\ReorderSectionsAction;
 use davidhirtz\yii2\skeleton\web\Controller;
 use Yii;
 use yii\filters\AccessControl;
@@ -211,7 +212,10 @@ class SectionController extends Controller
         $sectionIds = array_map('intval', array_filter(Yii::$app->getRequest()->post('section', [])));
 
         if ($sectionIds) {
-            $entry->updateSectionOrder($sectionIds);
+            Yii::createObject(ReorderSectionsAction::class, [
+                'entry' => $entry,
+                'sectionIds' => $sectionIds,
+            ]);
         }
     }