+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 2.0.2 (Nov 5, 2023)
 
+- Removed `ActiveRecord::updatePosition()`, `Category::updateEntryOrder()`, `Entry::updateAssetOrder()`,
+  `Entry::updateSectionOrder()`, `Section::updateAssetOrder()`, `Section::updateSectionOrder()` use model actions found
+  in `davidhirtz\yii2\cms\models\actions`
 - Added `Canvas` widget to display the assets in the frontend
 - Renamed `AssetViews` to `Gallery`
-- Renamed `SectionViews` to `Sections` 
+- Renamed `SectionViews` to `Sections`
 - Renamed `MetaTags::register()` to `MetaTags::widget()` to match the other widgets
 
 ## 2.0.1 (Nov 4, 2023)